backend/src/requirements.txt
Flask==2.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
pandas==1.5.3
requests==2.28.2
python-dotenv==1.0.0
pytz==2023.3